rivesaltes.git/requirements.txt
# Everything needed in production

pytest
h5py
boto3
boto
Pillow
whisper
cryptography
whisper
numpy
scipy
pvlib
tables
pandas #solar angles
deap #genetic algorithm
xlrd
nbconvert
matplotlib
sklearn
pyarrow
sphinx_bootstrap_theme
nbsphinx
m2r
sphinx_rtd_theme
sphinxcontrib-napoleon
